### 2026-01-05 34de628
Add Codecov badge to README
Added Codecov badge to README for test coverage.
