### 2026-01-19 #9668 @crvernon
:wave: @jbytecode would you like to edit this one?

Before it gets going, @zStupan will have to see to the new JOSS protocol mentioned [here](https://github.com/openjournals/joss-reviews/issues/9668#issuecomment-3760869673).
