### 2026-01-20 7828c21
feat: AI usage disclosure
