### 2026-01-20 5ad8fee
fix: AI usage disclosure
