### 2026-01-20 #9668 @HlisTilen
@jbytecode thank you for taking on the editor role. We have added an AI usage disclosure section to the manuscript, describing the limited use of language tools and clarifying that the core architecture and code were fully designed and implemented by the authors ([comment](https://github.com/openjournals/joss-reviews/issues/9668#issuecomment-3760869673)).
