### 2026-01-22 bf2ff39
feat: new Software Design section
