+++ paper/paper.md
@@ -26,9 +26,13 @@ bibliography: paper.bib
 
 NiaARM.jl is an open-source Julia package for Numerical Association Rule Mining (NARM) based on stochastic population-based nature-inspired optimization algorithms [@telikani2020survey]. It brings the capabilities of the original Python-based NiaARM framework [@stupan2022niaarm] to the Julia ecosystem, enabling researchers and data scientists working with datasets with mixed attribute types (consisting of categorical and numerical attributes) to discover numerical association rules. NiaARM.jl supports loading datasets, preprocessing, association rule mining, and extraction of discovered rules with associated interestingness metrics. In line with the rule mining part, this package also implements several well-known stochastic population-based nature-inspired algorithms, such as Differential Evolution (DE) [@storn1997differential], Artificial Bee Colony (ABC) [@karaboga2007powerful], Particle Swarm Optimization (PSO) [@kennedy1995particle], and several other metaphor-based nature-inspired algorithms to act as solvers for the numerical association rule mining task. The entire numerical association rule mining workflow is further supported by visualization methods for numerical association rules, which is achieved through NarmViz.jl, a package well integrated with NiaARM.jl [@fister2024narmviz].
 
-The flow of the NiaARM.jl framework is shown in \autoref{fig:NiaARM}. Users can construct a dataset either from a CSV file (via a file path) or directly from `DataFrame`. The dataset is then used to build the numerical association rule mining optimization problem together with user-selected interestingness measures, which are used in the computation of the fitness function. The optimization problem can be solved using any of the population-based nature-inspired algorithms implemented in NiaARM.jl (e.g., DE, ABC, PSO) to mine numerical association rules from the dataset. The discovered rules are returned as a collection of `Rule` objects and can be further analyzed, exported (e.g., to CSV in downstream post-processing), or visualized using NarmViz.jl [@fister2024narmviz], which provides several visualization methods for numerical association rules.
+# Software Design
+
+The core of NiaARM.jl is organized around four main modules: (1) the `Dataset` module, which handles loading and preprocessing of transaction data from CSV files or DataFrames, representing numerical attributes as interval features and categorical attributes as sets of categories; (2) the `Problem` module, which formulates numerical association rule mining as a continuous optimization problem, where candidate solutions are encoded as real-valued vectors and decoded into association rules; (3) the optimization algorithms module, which provides implementations of various nature-inspired algorithms that explore the search space to identify rules; and (4) the `Rule` module, which represents discovered association rules and provides methods for computing interestingness measures.
 
 ![NiaARM.jl flow.\label{fig:NiaARM}](NiaARM_jl_Pipeline.png)
+
+The flow of the NiaARM.jl framework is shown in \autoref{fig:NiaARM}. Users can construct a dataset either from a CSV file (via a file path) or directly from `DataFrame`. The dataset is then used to build the numerical association rule mining optimization problem together with user-selected interestingness measures, which are used in the computation of the fitness function. The optimization problem can be solved using any of the population-based nature-inspired algorithms implemented in NiaARM.jl (e.g., DE, ABC, PSO) to mine numerical association rules from the dataset. The discovered rules are returned as a collection of `Rule` objects and can be further analyzed, exported (e.g., to CSV in downstream post-processing), or visualized using NarmViz.jl [@fister2024narmviz], which provides several visualization methods for numerical association rules.
 
 # Statement of need
 
