### 2026-01-23 8b18f51
feat: Research impact statement
