### 2026-01-23 ac70f7f
fix: small documentation corections
