### 2026-01-26 13c32d1
Fix paper.md
