+++ paper/paper.md
@@ -48,11 +48,9 @@ Despite the growing importance of NARM in domains such as finance, sport, and me
 
 # Research impact statement
 
-The application of nature-inspired algorithms to numerical association rule mining has been explored in several studies [@fister2018differential], leading to the development of the Python based NiaARM framework [@stupan2022niaarm]. These tools have been applied in various domains, including sport science, healthcare, smart agriculture, and environmental monitoring, demonstrating the versatility of the optimization-driven approach to rule mining.
+NiaARM.jl framework lowers the barrier to high-performance NARM by making NARM accessible, faster than previous implementations especially on complex datasets, and extensible within the Julia ecosystem. By providing a dedicated Julia framework for dealing with NARM tasks, the package extends an established methodology that was previously available primarily in the Python programming language. It allows researchers and practitioners who rely on Julia for high-performance scientific computing to solve NARM tasks. NiaARM.jl is also well integrated with other packages, for example NarmViz, which extends the rule mining pipeline to include visualization.
 
-The impact of NiaARM.jl lies in providing the dedicated Julia framework for numerical association rule mining and also in extending the reach of this established methodology to a new community of researchers who benefit from Julia's strengths in high-performance scientific computing. By integrating with NarmViz.jl [@fister2024narmviz] for visualization, NiaARM.jl offers visual representations of discovered rules. This contributes to greater trust and usability of mined rules in sensitive domains such as healthcare and environmental sciences, or in general where interpretability is important.
-
-Furthermore, NiaARM.jl provides implementations of multiple nature-inspired algorithms that can serve as reference implementations for researchers studying metaheuristic optimization, thereby contributing to reproducibility and comparative studies in the field. The modular architecture encourages community-driven extensions, enabling researchers to rapidly prototype new algorithmic variants and interestingness measures.
+In addition to its applied impact, NiaARM.jl contributes to methodological research by providing reference implementations of multiple stochastic population-based nature-inspired algorithms within a unified framework. This supports reproducibility, benchmarking, and comparative studies in metaheuristic optimization beyond solving NARM tasks alone. The modular design of the framework encourages community-driven extensions, enabling rapid prototyping of new algorithms, fitness functions, and interestingness measures.
 
 ## AI usage disclosure
 
