### 2026-01-26 #9668 @HlisTilen
@jbytecode, we have added `Software Design` and `Research Impact Statement` sections in the paper. Thank you!
