### 2026-01-26 #9668 @jbytecode
@zStupan - Do you suggest any potential reviewers for this submission? You can use the list of people link (https://reviewers.joss.theoj.org/lookup) to search people by their expertise and programming languages. Please do not use `@` character to prevent unwanted notifications. Thank you in advance.
