### 2026-01-28 #9668 @jbytecode
@zStupan - As our editor-in-chief states, we need to add one more section. The required sections for JOSS articles are currently being restructured, and since this process is very recent, it has caused some inconsistencies. Please accept our apologies for the inconvenience. Additionally, if you have any suggestions for further reviewers, please let us know; otherwise, I will look into finding suitable candidates for you. Thank you in advance.
