### 2026-01-28 #9668 @HlisTilen
@jbytecode thank you for the clarification and for the additional details. No problem at all, we will prepare the requested section (State of the Field) within the next 2-3 days and update the manuscript accordingly.  Regarding additional reviewers, we would like to suggest Erkan Karabulut ([https://github.com/erkankarabulut](https://github.com/erkankarabulut)). He works in a closely related research area and has strong expertise that aligns well with the topic and scope of our software, so we believe he would be a very suitable reviewer.
