### 2026-01-29 bcfe04e
feat: State of the Field
