+++ paper/paper.bib
@@ -86,3 +86,45 @@
   organization={Springer}
 }
 
+@article{hahsler2005arules,
+  title   = {arules - A Computational Environment for Mining Association Rules and Frequent Item Sets},
+  author  = {Hahsler, Michael and Gr{\"u}n, Bettina and Hornik, Kurt},
+  journal = {Journal of Statistical Software},
+  volume  = {14},
+  number  = {15},
+  pages   = {1--25},
+  year    = {2005},
+  doi     = {10.18637/jss.v014.i15}
+}
+
+@article{fournier2020spmf,
+  title={Spmf: A java open-source data mining library},
+  author={Fournier-Viger, Philippe and others},
+  journal={URL: www. philippe-fournier-viger. com/spmf/(visited on 08/15/2018)},
+  year={2020}
+}
+
+@misc{ruleminerjl,
+  title        = {RuleMiner.jl: Efficient Frequent Itemset and Association Rule Mining in Julia},
+  author       = {Schwartz, Jared},
+  howpublished = {\url{https://github.com/JaredSchwartz/RuleMiner.jl}},
+  year         = {2024},
+  note         = {Accessed: 2026-01-29}
+}
+
+@article{fister2020uarmsolver,
+  title={uarmsolver: A framework for association rule mining},
+  author={Fister, Iztok and Fister Jr, Iztok},
+  journal={arXiv preprint arXiv:2010.10884},
+  year={2020}
+}
+
+@article{fister2026niarules,
+  title={niarules: Advancing interpretable machine learning through numerical association rule mining and 3D coral plot visualization},
+  author={Fister Jr, Iztok and Emsenhuber, Gerlinde and Pl{\"u}mer, Jan Hendrik and Fister, Iztok and Holzinger, Andreas},
+  journal={SoftwareX},
+  volume={33},
+  pages={102470},
+  year={2026},
+  publisher={Elsevier}
+}
