### 2026-01-29 #9668 @jbytecode
👋 @erkankarabulut - Would you be willing to review this submission to JOSS? You are suggested by the author(s).

We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines:

https://joss.readthedocs.io/en/latest/review_criteria.html
