### 2026-02-02 #9668 @Kiwy3
Hi @jbytecode  — yes, I can review this submission for JOSS. Thanks for the invitation.
