### 2026-02-02 #9925 @editorialbot
[REVIEW]: NiaARM.jl: A Julia Framework for Numerical Association Rule Mining Using Nature‑Inspired Optimization Algorithms
**Submitting author:** @zStupan (<a href="http://orcid.org/0000-0001-9847-7306">Žiga Stupan</a>)
**Repository:** https://github.com/firefly-cpp/NiaARM.jl
**Branch with paper.md** (empty if default branch):
**Version:** 0.3.2
