+++ paper/paper.bib
@@ -83,7 +83,8 @@
   booktitle={International conference on intelligent data engineering and automated learning},
   pages={79--88},
   year={2018},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-030-03493-1_9}
 }
 
 @article{hahsler2005arules,
@@ -101,7 +102,8 @@
   title={Spmf: A java open-source data mining library},
   author={Fournier-Viger, Philippe and others},
   journal={URL: www. philippe-fournier-viger. com/spmf/(visited on 08/15/2018)},
-  year={2020}
+  year={2020},
+  doi={10.5555/2627435.2750353}
 }
 
 @misc{ruleminerjl,
@@ -116,7 +118,8 @@
   title={uarmsolver: A framework for association rule mining},
   author={Fister, Iztok and Fister Jr, Iztok},
   journal={arXiv preprint arXiv:2010.10884},
-  year={2020}
+  year={2020},
+  doi={10.48550/arXiv.2010.10884}
 }
 
 @article{fister2026niarules,
@@ -126,5 +129,6 @@
   volume={33},
   pages={102470},
   year={2026},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.softx.2025.102470}
 }
