### 2026-02-16 #9925 @Kiwy3
Hello,

I’m not sure if this is the appropriate place to raise this, but it appears that the documentation for NarmViz.jl is not compatible with the current version of NiaARM.jl. Since it is referenced in the paper as part of the NiaARM workflow, I would expect it to function correctly.

Aside from this point, I have completed my review. As a minor comment, I would note that there currently seems to be limited external engagement, which raises some questions about the collaborative aspect of the repository.
