### 2026-02-18 #9925 @zStupan
Hello @Kiwy3,

Thank you for your review. We have updated the NarmViz.jl documentation to reflect the changes in NiaARM.jl.
