### 2026-02-19 #9925 @LaurenzBeck
I have a few questions/topics to clarify:

1. the **open development** section asks for a six month development timeline. the repository was created on the 15.12.2025, so its only three months of development. however, since many components/algorithms were ported from the other Niarm packages from the authors, I suspect that we can attribute some of the development time of those packages for the publication.
2. regarding **performance**: the paper mentions the speed of Julia as a motivation for porting the library to the new ecosystem. Given the very similar API and design of NiaPy, it would be interesting to compare the runtime and memory consumption of a few benchmarks/settings.
3. I like the paper, but I think it would benefit from a short introduction into association rule mining.

Other than that, the repository and codebase look good to me 💪🔥

have you considered my comments from the pre-review thread? https://github.com/openjournals/joss-reviews/issues/9668#issuecomment-3705346656
