### 2026-02-19 #9925 @firefly-cpp
Dear @LaurenzBeck,

Thank you very much for your review!

1. The repository is almost four years old, and development began on May 7 2022. Please take a look here: https://github.com/firefly-cpp/NiaARM.jl/commits/main/?after=1f7340329d0d21ed519be3e359424616cf14ee46+174

2. We will also consider appropriate benchmarks for a fair comparison.

3. Good point! We will add the suggested additional content to the paper.

@zStupan has already incorporated the comments from your pre‑review.

Thanks again for your valuable feedback.
