### 2026-03-03 #9925 @HlisTilen
@jbytecode Thank you for the reminder.

We have already addressed the comment regarding the need for a short introduction to association rule mining and added a concise introductory section to improve clarity for readers who may not be familiar with the topic.

Regarding the performance comparison (runtime and memory consumption benchmarks against NiaPy), we are currently working on this analysis and expect to finalize the results in the coming days.
