### 2026-03-06 #9925 @HlisTilen
@LaurenzBeck

Thank you for the constructive feedback and for the helpful suggestions in the pre-review discussion.

Regarding the performance motivation mentioned in the paper, we conducted a small benchmark comparing the runtime of the original Python implementation (NiaARM) and the Julia implementation (NiaARM.jl) under comparable settings. The benchmark evaluated Differential Evolution (DE) and Particle Swarm Optimization (PSO) with 1,000 and 10,000 iterations.

The results indicate a consistent runtime improvement when using the Julia implementation. For example, in the DE configuration with 1,000 iterations, the Python version required 33.3 s, while NiaARM.jl completed the same task in 8.2 s. Similarly, for PSO with 1,000 iterations, the runtime decreased from 15.8 s to 2.4 s. In larger runs (10,000 iterations), the Julia implementation also demonstrated significant speedups, reducing runtime from 1091.5 s to 403.1 s for DE and from 140.0 s to 11.4 s for PSO. These results support the motivation for porting the framework to Julia and highlight the performance benefits of the language in computationally intensive optimization tasks.

However, we decided not to include this benchmark comparison in the JOSS paper itself. Since JOSS papers are intentionally concise and focused primarily on the software contribution, a detailed performance comparison would require additional explanation, experimental setup description, and discussion that would take a significant portion of the limited article space and potentially distract from the main goal of the paper namely presenting the design and capabilities of the NiaARM.jl framework.

For transparency and completeness, we are therefore posting the benchmark results here in the review thread instead.

We appreciate the suggestion and believe that a more extensive benchmark study comparing implementations across ecosystems would be valuable in a separate methodological or experimental publication.
