### 2026-03-12 #9925 @LaurenzBeck
Very nice benchmark, thank you for putting in the extra effort!

last time I checked other JOSS papers (around 2024), the 2 page limit was not strict and many authors included such extra details.
I can understand, however, that the paper length is enforced now and you want to keep it concise.

I would recommend to additionally put the paragraph and figure in the README or the hosted docs.

I had a look at the new section interation on association rule mining (https://github.com/firefly-cpp/NiaARM.jl/blob/main/paper/paper.md#state-of-the-field) and think its a good introduction given the space constraints, good work 💪

Now that these two concerns of mine are adressed, I can recommend the paper to be accepted :)
