### 2026-03-12 #9925 @HlisTilen
@LaurenzBeck Thank you for the kind feedback and for taking the time to review the updated version. Following your recommendation, we will include the corresponding paragraph and figure in the project README.

Thank you again for your constructive comments and support throughout the review process. We appreciate your recommendation for acceptance.
