### 2026-03-27 #9925 @jbytecode
@zStupan - I've just sent a pull request:

https://github.com/firefly-cpp/NiaARM.jl/pull/54

This PR includes minor fixes in both the manuscript and the bibliography. Please review and merge. Thank you.
