### 2026-03-27 #9925 @jbytecode
@zStupan - Thank you, the paper and the bibliography look better now.

Please perform these tasks:

- Create a tagged release in the software repository, e.g. `v1.2.3`. This tagged release is a proof of reviewed code.
- Use this tag to create an archive in the Zenodo. In the Zenodo, please set the software license properly (e.g. MIT)
- The author list, their orcid, and affilications in the archive should match those in the paper.
- The title of the Zenodo archive must be the same of the paper's title.

Please report me the tagged release, the arcive URL, and the Zenodo DOI.

Thank you in advance.
