+++ 0.3.1/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"api_reference/#API-Reference","page":"API Reference","title":"API Reference","text":"","category":"section"},{"location":"api_reference/#Data-and-features","page":"API Reference","title":"Data and features","text":"","category":"section"},{"location":"api_reference/#Attributes-and-rules","page":"API Reference","title":"Attributes and rules","text":"","category":"section"},{"location":"api_reference/#Metrics","page":"API Reference","title":"Metrics","text":"","category":"section"},{"location":"api_reference/#Mining-pipeline","page":"API Reference","title":"Mining pipeline","text":"","category":"section"},{"location":"api_reference/#Optimization-setup","page":"API Reference","title":"Optimization setup","text":"","category":"section"},{"location":"api_reference/#Algorithms","page":"API Reference","title":"Algorithms","text":"","category":"section"},{"location":"api_reference/#Utilities","page":"API Reference","title":"Utilities","text":"","category":"section"},{"location":"api_reference/#NiaARM.Dataset","page":"API Reference","title":"NiaARM.Dataset","text":"Dataset(df_or_path)\n\nWrap a tabular dataset and derive feature metadata used by optimizers. Accepts either a DataFrame or path to a CSV file. Features are inferred from column types and overall problem dimensionality is computed for use with optimization algorithms.\n\n\n\n\n\n","category":"type"},{"location":"api_reference/#NiaARM.AbstractFeature","page":"API Reference","title":"NiaARM.AbstractFeature","text":"AbstractFeature\n\nAbstract supertype describing dataset columns used to build candidate rules.\n\n\n\n\n\n","category":"type"},{"location":"api_reference/#NiaARM.NumericalFeature","page":"API Reference","title":"NiaARM.NumericalFeature","text":"NumericalFeature(name, min, max)\n\nMetadata describing a numerical column along with its observed minimum and maximum. Bounds are used to scale optimizer solutions back into meaningful ranges.\n\n\n\n\n\n","category":"type"},{"location":"api_reference/#NiaARM.CategoricalFeature","page":"API Reference","title":"NiaARM.CategoricalFeature","text":"CategoricalFeature(name, categories)\n\nMetadata describing a categorical column and its possible category values.\n\n\n\n\n\n","category":"type"},{"location":"api_reference/#NiaARM.dtype","page":"API Reference","title":"NiaARM.dtype","text":"dtype(attribute)\n\nReturn the underlying numeric element type carried by an attribute or feature. For categorical elements, String is returned.\n\n\n\n\n\n","category":"function"},{"location":"api_reference/#NiaARM.isnumerical","page":"API Reference","title":"NiaARM.isnumerical","text":"isnumerical(feature)\n\nReturn true when a feature is numerical.\n\n\n\n\n\nisnumerical(attribute)\n\nReturn true when an attribute is numerical.\n\n\n\n\n\n","category":"function"},{"location":"api_reference/#NiaARM.iscategorical","page":"API Reference","title":"NiaARM.iscategorical","text":"iscategorical(feature)\n\nReturn true when a feature is categorical.\n\n\n\n\n\niscategorical(attribute)\n\nReturn true when an attribute is categorical.\n\n\n\n\n\n","category":"function"},{"location":"api_reference/#NiaARM.AbstractAttribute","page":"API Reference","title":"NiaARM.AbstractAttribute","text":"AbstractAttribute\n\nAbstract supertype for all attribute descriptors that can appear in a mined rule.\n\n\n\n\n\n","category":"type"},{"location"
[…]
