+++ 0.3.1/siteinfo.js
@@ -1,0 +1,2 @@
+var DOCUMENTER_CURRENT_VERSION = "0.3.1";
+var DOCUMENTER_IS_DEV_VERSION = false;
