+++ versions.js
@@ -3,5 +3,5 @@ var DOC_VERSIONS = [
   "v0.3",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "0.3.1";
+var DOCUMENTER_NEWEST = "0.3.2";
 var DOCUMENTER_STABLE = "stable";
