### 2026-03-30 #9925 @HlisTilen
@jbytecode  thank you for your guidance.

We have completed the requested steps:

**Tagged release**: https://github.com/firefly-cpp/NiaARM.jl/releases/tag/0.3.2
**Zenodo archive**: https://zenodo.org/records/19331323
**Zenodo DOI**: 10.5281/zenodo.19331323

The Zenodo record has been updated to ensure that the title, author list, ORCID identifiers, affiliations, and software license are fully aligned with the paper.

Please let us know if any further adjustments are needed.
