### 2026-03-30 #9925 @jbytecode
@HlisTilen , @zStupan - The whole stuff looks better than ever. Please, perform a full proof-read of the paper. Fix any typos and minor things if exist. Please tell me your change log (list of commits) and ping me when you've done with it. Thank you in advance.
