### 2026-03-31 #9925 @HlisTilen
@jbytecode, thank you for your feedback. We appreciated it a lot.

We have performed a full proofread of the paper and corrected minor typos.

Change log:

- fix: correct punctuation in NARM pipeline description and research impact statement

The changes are reflected in the following commit:

https://github.com/firefly-cpp/NiaARM.jl/commit/55623c73266e57d64b4fa9dcd3e7db95b46c6eff

No major changes were introduced. Please let us know if any further adjustments are needed.
